spreadsheet-generator/datasheet_d0w/requirements.txt
pandas
psycopg2-binary
sqlalchemy
numpy
requests
dotenv
